Changes to enable building all MacOS types on any MacOS host
This makes changes to enable us to build for all supported
MacOS versions on any kind of MacOS host running the latest version.
Most of the actual changes are in puppet-runtime, as things in this repo
are mostly around packaging. But to be safe, we use the same strategy as
we do in puppet-runtime by running the build command in a shell invoked
via the 'arch' command to support building x86_64 on an arm64 host.
Because we now build a single agent for all supported MacOS versions, we
use osx-15 as the platform string and remove the others.